+++ kernel/include/kern/signal.h
@@ -1,7 +1,6 @@
 #ifndef _KERN_SIGNAL_H_
 #define _KERN_SIGNAL_H_
 
-#include "kern/types.h"
 #include <signal.h>
 
                       /* SIG */
@@ -24,9 +23,9 @@
 
 struct signal_t
 {
-        uchar allocated; /* Whether or not this entry is allocated. */
-        uchar default_action; /* Default action */
-        uchar catchable; /* Can this signal be caught? */
+        int allocated; /* Whether or not this entry is allocated. */
+        int default_action; /* Default action */
+        int catchable; /* Can this signal be caught? */
         int signum; /* The number of the signal */
 
         /**
