+++ kernel/ioctl.c
@@ -1,10 +1,11 @@
-#include "kern/types.h"
+#include <stdint.h>
+#include <stdlib.h>
 #include "syscall.h"
 
 /**
  * ioctl helper function
  */
-uchar ioctl_arg_ok(void* addr, uint sz)
+int ioctl_arg_ok(void* addr, size_t sz)
 {
 	if(syscall_addr_safe(addr) || syscall_addr_safe(addr + sz - 1))
                 return 1;
